processor/sinkmanager(ticdc): improve code struct and remove useless abstraction
