+++ cdc/processor/sinkmanager/manager_test.go
